requirements.txt
astroid==1.6.5
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==1.0.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
marshmallow==2.15.3
mccabe==0.6.1
pylint==1.9.2
PyYAML==3.13
requests==2.19.1
six==1.11.0
slackclient==1.2.1
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
